Move protocol messages from edb/testbase/protocol to edb/protocol
